rutas-protegidas-api/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
python-dotenv==1.0.0